// src/app/models/forecast.model.ts
export class forecast {
    city: String;
    currentDay1: String;
    currentDay2: String;
    currentDay3: String;
    currentDay4: String;
    currentDay5: String;
    day1temp: number;
    day2temp: number;
    day3temp: number;
    day4temp: number;
    day5temp: number;
    day1icon: String;
    day2icon: String;
    day3icon: String;
    day4icon: String;
    day5icon: String;
  }